requirements.txt
pytorch-msssim
torch
torchvision
Pillow
opencv-python